Share platform-split module types via .shared.ts files
Extract the prop/control types (and the video error classes) that were
duplicated between web implementations and their native throw-stubs into
sibling .shared.ts files, so both variants reference one definition and
cannot drift.
