+++ src/components/Post/Embed/VideoEmbed/VideoEmbedInner/VideoEmbedInnerWeb.native.tsx
@@ -1,23 +1,10 @@
-import {type AppBskyEmbedVideo} from '@atproto/api'
+import {type VideoEmbedInnerWebProps} from './VideoEmbedInnerWeb.shared'
 
-export function VideoEmbedInnerWeb(_props: {
-  embed: AppBskyEmbedVideo.View
-  active: boolean
-  setActive: () => void
-  onScreen: boolean
-  lastKnownTime: React.RefObject<number | undefined>
-}): never {
+export {
+  HLSUnsupportedError,
+  VideoNotFoundError,
+} from './VideoEmbedInnerWeb.shared'
+
+export function VideoEmbedInnerWeb(_props: VideoEmbedInnerWebProps): never {
   throw new Error('VideoEmbedInnerWeb may not be used on native.')
-}
-
-export class HLSUnsupportedError extends Error {
-  constructor() {
-    super('HLS is not supported')
-  }
-}
-
-export class VideoNotFoundError extends Error {
-  constructor() {
-    super('Video not found')
-  }
 }
