+++ src/components/Post/Embed/VideoEmbed/VideoEmbedInner/VideoEmbedInnerWeb.tsx
@@ -1,6 +1,5 @@
 import {useCallback, useEffect, useId, useRef, useState} from 'react'
 import {View} from 'react-native'
-import {type AppBskyEmbedVideo} from '@atproto/api'
 import {msg} from '@lingui/core/macro'
 import {useLingui} from '@lingui/react'
 import type * as HlsTypes from 'hls.js'
@@ -10,7 +9,17 @@ import {atoms as a} from '#/alf'
 import {AltBadgeWithDialog} from '#/components/AltBadgeWithDialog'
 import {useFullscreen} from '#/components/hooks/useFullscreen'
 import * as BandwidthEstimate from './bandwidth-estimate'
+import {
+  HLSUnsupportedError,
+  type VideoEmbedInnerWebProps,
+  VideoNotFoundError,
+} from './VideoEmbedInnerWeb.shared'
 import {Controls} from './web-controls/VideoControls'
+
+export {
+  HLSUnsupportedError,
+  VideoNotFoundError,
+} from './VideoEmbedInnerWeb.shared'
 
 export function VideoEmbedInnerWeb({
   embed,
@@ -18,13 +27,7 @@ export function VideoEmbedInnerWeb({
   setActive,
   onScreen,
   lastKnownTime,
-}: {
-  embed: AppBskyEmbedVideo.View
-  active: boolean
-  setActive: () => void
-  onScreen: boolean
-  lastKnownTime: React.RefObject<number | undefined>
-}) {
+}: VideoEmbedInnerWebProps) {
   const containerRef = useRef<HTMLDivElement>(null)
   const videoRef = useRef<HTMLVideoElement>(null)
   const [focused, setFocused] = useState(false)
@@ -104,12 +107,6 @@ export function VideoEmbedInnerWeb({
   )
 }
 
-export class HLSUnsupportedError extends Error {
-  constructor() {
-    super('HLS is not supported')
-  }
-}
-
 // Bluesky serves HLS as MPEG-TS with H.264 + AAC. `Hls.isSupported()` is loose
 // (true if MSE supports *any* of {H.264, AV1, VP9} OR *any* of {AAC, FLAC}),
 // so on Linux boxes missing H.264 (e.g. no ubuntu-restricted-extras, sandboxed
@@ -136,12 +133,6 @@ function canPlayBskyVideoCodecs(): boolean {
     mediaSource.isTypeSupported('video/mp4; codecs="avc1.42E01E"') &&
     mediaSource.isTypeSupported('audio/mp4; codecs="mp4a.40.2"')
   )
-}
-
-export class VideoNotFoundError extends Error {
-  constructor() {
-    super('Video not found')
-  }
 }
 
 type CachedPromise<T> = Promise<T> & {value: undefined | T}
