+++ src/components/Post/Embed/VideoEmbed/VideoEmbedInner/web-controls/VideoControls.native.tsx
@@ -1,19 +1,5 @@
-import type Hls from 'hls.js'
+import {type ControlsProps} from './VideoControls.shared'
 
-export function Controls(_props: {
-  videoRef: React.RefObject<HTMLVideoElement | null>
-  hlsRef: React.RefObject<Hls | undefined | null>
-  active: boolean
-  setActive: () => void
-  focused: boolean
-  setFocused: (focused: boolean) => void
-  onScreen: boolean
-  fullscreenRef: React.RefObject<HTMLDivElement | null>
-  hlsLoading: boolean
-  hasSubtitleTrack: boolean
-  isGif: boolean
-  altText?: string
-  updateCuePositions: (controlsVisible?: boolean) => void
-}): never {
+export function Controls(_props: ControlsProps): never {
   throw new Error('VideoWebControls may not be used on native.')
 }
