+++ src/components/hooks/useWelcomeModal.native.ts
@@ -1,7 +1,5 @@
-export function useWelcomeModal(): {
-  isOpen: boolean
-  open: () => void
-  close: () => void
-} {
+import {type WelcomeModalControl} from './useWelcomeModal.shared'
+
+export function useWelcomeModal(): WelcomeModalControl {
   throw new Error('useWelcomeModal is web only')
 }
