+++ src/components/hooks/useWelcomeModal.shared.ts
@@ -1,0 +1,5 @@
+export type WelcomeModalControl = {
+  isOpen: boolean
+  open: () => void
+  close: () => void
+}
