+++ src/view/com/composer/videos/SubtitleFilePicker.native.tsx
@@ -1,6 +1,5 @@
-export function SubtitleFilePicker(_props: {
-  onSelectFile: (file: File) => void
-  disabled?: boolean
-}): never {
+import {type SubtitleFilePickerProps} from './SubtitleFilePicker.shared'
+
+export function SubtitleFilePicker(_props: SubtitleFilePickerProps): never {
   throw new Error('SubtitleFilePicker is a web-only component')
 }
