+++ src/view/com/composer/videos/SubtitleFilePicker.shared.ts
@@ -1,0 +1,4 @@
+export type SubtitleFilePickerProps = {
+  onSelectFile: (file: File) => void
+  disabled?: boolean
+}
